LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-fit/line-height-expected.txt

xxxx
xxxx
xxxx
xxxx
xxxx
xxxx
xxxx
xxxx

FAIL line-height:normal should be scaled assert_greater_than: expected a number greater than 18 but got 18
FAIL line-height:<number> should be scaled assert_greater_than: expected a number greater than 26 but got 26
PASS line-height:<percentage> should not be scaled
PASS line-height:<length> should not be scaled
